refactor(validation): modularize implementation-fidelity-validation.js (1,559 → 10 modules)
SD-LEO-REFACTOR-IMPL-FIDELITY-001 - Child 8 of orchestrator SD-LEO-REFACTOR-LARGE-FILES-002

Extracted 1,559 LOC monolith into focused modules:

Utils (3 files):
- git-helpers.js - SD search terms, git log functions
- repo-detection.js - Implementation repo detection (EHG vs EHG_Engineer)
- index.js - Utility exports

Preflight (1 file):
- index.js - All 4 preflight checks (app directory, ambiguity, server restart, stubbed code)

Sections (5 files):
- design-fidelity.js - Section A validation (25 points)
- database-fidelity.js - Section B validation (35 points)
- data-flow-alignment.js - Section C validation (25 points)
- enhanced-testing.js - Section D validation (25 points)
- index.js - Section exports

Main (1 file):
- index.js - Main orchestrator with validateGate2ExecToPlan

Original file now thin re-export (20 LOC). 12 exports verified.
