+++ scripts/modules/implementation-fidelity/sections/index.js
@@ -1,0 +1,9 @@
+/**
+ * Section Exports for Implementation Fidelity Validation
+ * Part of SD-LEO-REFACTOR-IMPL-FIDELITY-001
+ */
+
+export { validateDesignFidelity } from './design-fidelity.js';
+export { validateDatabaseFidelity } from './database-fidelity.js';
+export { validateDataFlowAlignment } from './data-flow-alignment.js';
+export { validateEnhancedTesting } from './enhanced-testing.js';
