+++ scripts/modules/implementation-fidelity/utils/index.js
@@ -1,0 +1,16 @@
+/**
+ * Utility Exports for Implementation Fidelity Validation
+ * Part of SD-LEO-REFACTOR-IMPL-FIDELITY-001
+ */
+
+export {
+  getSDSearchTerms,
+  gitLogForSD,
+  clearSearchTermsCache
+} from './git-helpers.js';
+
+export {
+  detectImplementationRepo,
+  EHG_ENGINEER_ROOT,
+  EHG_ROOT
+} from './repo-detection.js';
